Splitting out more server handlers for easier testing
