+++ block-streamer/src/main.rs
@@ -1,3 +1,4 @@
+mod handlers;
 mod receive;
 mod receiver;
 mod server;
